[models] Turns out, don't need to isolate_namespace Spree
Calling isolate_namespace in models was telling Rails to use Spree::Models (which has no routes defined), for the spree.* routes. Views for spree were being evaluated within the "routing context" of Spree::Models, not Spree::Core. This means to get to the actual routes for core, routing helpers needed to have a spree.* prefix.

Remove the isolate_namespace call and the problem goes away.